fix: Merge main.js and main.ts + fix search routes redirection - Remove duplicate main.js and consolidate into main.ts - Add redirection from /search to /videos for history
